Adding support for DiaNavigation in UWP test adapter (#258)
* Adding support for DiaNavigation in UWP test adapter

Testing
Verifed that for UWP test running via Tpv1, n exception is thrown

Once we take dependency on Tpv2 Obj Model, we can enable commented testcases

* Making Dia operation common for netcore, & uwp

* move initialization of DiaSessionOperations
