+++ src/Adapter/PlatformServices.Portable/PlatformServices.Portable.csproj
@@ -43,6 +43,9 @@
     <Compile Include="..\PlatformServices.Shared\netstandard1.0\Services\ns10FileOperations.cs">
       <Link>Services\ns10FileOperations.cs</Link>
     </Compile>
+	<Compile Include="..\PlatformServices.Shared\netstandard1.0\Services\ns10DiaSessionOperations.cs">
+      <Link>Services\ns10DiaSessionOperations.cs</Link>
+    </Compile>
     <Compile Include="..\PlatformServices.Shared\netstandard1.0\Services\ns10ReflectionOperations.cs">
       <Link>Services\ns10ReflectionOperations.cs</Link>
     </Compile>
